requirements.txt
pymupdf
requests
python-dotenv
sentence-transformers
langchain-community
faiss-cpu
fpdf
